Switch to -Include instead of -Filter to fix PWSH issue
I was finding that Install-Module PSGitHub worked but that many of the functions weren't working. I tracked it down to the fact that the Get-ChildItem call here wasn't recursing into the Functions directory. I think that this may be a behavior change for the -Filter parameter in PSCore, but -Include works across both PS5 and PSCore 6 so I am proposing to change.